+++ CHANGELOG.md
@@ -23,6 +23,10 @@
 - :game_die: Add support for subtraction of expressions (e.g. `d20 - 5`)  
   [Jonathan Hoffman](https://github.com/JonathanHoffman) [#80](https://github.com/TabletopAssistant/DiceKit/pull/80)
 
+- :game_die: Add min/max operations.
+  [Logan Johnson](https://github.com/loganjohnson)
+  [#64](https://github.com/TabletopAssistant/DiceKit/pull/63)
+
 ## 0.1: It Starts with a Single Die (2015-07-25)
 
 - :game_die: Support rolling individual dice.